+++ inst/shiny-examples/BCGcalc/global.R
@@ -1,7 +1,7 @@
 # Shiny Global File
 
 # Version ----
-pkg_version <- "2.0.0.9005"
+pkg_version <- "2.0.0.9006"
 
 # Packages----
 library(BCGcalc)